kobe2box

Creates summary data.frame for presenting advice using VPA2Box projections

# S4 method for character
kobe2box(path, proxy = c("fmsy", "fmax", "f0.1", "f20",
  "f30", "f40", "f90max", "f75max")[3], what = c("sims", "trks", "pts",
  "smry", "wrms")[1], prob = c(0.75, 0.5, 0.25), year = NULL, nwrms = 10)

Arguments

path

character directory or directories with target files

proxy

character proxy for msy "f0.1" by default otherwise one of c("fmsy","fmax","f0.1","f20","f30","f40","f90max","f75max")

what

character output data.frame all "sims" by default or one of c("sims","trks","pts","smry","wrms")[1]

prob

numeric with probabilities, by default c(0.75,0.5,0.25)

year

numeric for year when slecting points

nwrms

numeric number of worms to select by default 10

Value

a data.frame or a list of data.frame with either all the simulations from PRO2BOX, i.e. for both historical assessment results and projections. These can either be in the form of summaries or all reults

Examples

## Not run: ------------------------------------
# sims=kobe2box("vpa")
## ---------------------------------------------