Kobe Phase Plot with Marginal Densities

Plots a Kobe Phase Plot with marginal densities '

kobePhaseMar(pts, trks = NULL, mns = FALSE, size = 1,
  xlab = expression(B:B[MSY]), ylab = expression(F:F[MSY]), xlim = 2,
  ylim = xlim, col = colorRampPalette(c("orange", "blue"), space = "Lab"),
  shade = 0.5, col2 = grey(shade), col3 = grey(shade * 1.1),
  layer = NULL)

Arguments

pts

data.frame with stock and harvest variables respresenting points for a year

trks

data.frame with stock and harvest variables respresenting tracks, NULL by default

mns

logical, do you want mean by year of pts be plotted, FALSE by default

size

of points1 by default

xlab

character or expression for x axis title, expression(B:B[MSY]) by default

ylab

character or expression for x axis title, expression(F:F[MSY]) by default

xlim

upper limit of x axis, 2 by default

ylim

upper limit of x axis, 2 by default

col

colorRampPalette(c("orange","blue"),space="Lab") by default

shade

.5 by default

col2

grey(shade) by default

col3

grey(shade*1.1) by default

layer

a ggplot2 layer to add to phase plot

Value

a list with 3 ggplot objects, the 2 marginal densities and the phase plot

Examples

   ## Not run: rnorm(10)